code Restructured | all views moved to views folder
