+++ backend/__init__.py
@@ -22,8 +22,8 @@ def create_app():
 
     db.init_app(app)
 
-    from backend.auth import auth 
-    from backend.views import views
+    from backend.views.auth import auth 
+    from backend.views.views import views
 
     app.register_blueprint(views,url_prefix='/')
     app.register_blueprint(auth,url_prefix='/auth/')
